+++ antarest/output/storage/output_storage.py
@@ -25,7 +25,7 @@ from pydantic.alias_generators import to_camel
 from antarest.core.serde import AntaresBaseModel
 from antarest.launcher.adapters.abstractlauncher import SimulationLogs
 from antarest.launcher.model import LogType
-from antarest.output.filestudy.utils import QueryFileType
+from antarest.output.filestudy.model import QueryFileType
 from antarest.output.model import OutputVariablesList
 from antarest.output.model.download import MatrixIndex
 from antarest.study.business.model.config.general_model import Mode
