+++ antarest/output/storage/v2/storage.py
@@ -39,11 +39,11 @@ from antarest.core.utils.utils import StopWatch
 from antarest.launcher.adapters.abstractlauncher import SimulationLogs
 from antarest.launcher.model import LogType
 from antarest.lfs.lfs import ILargeFileStorage
-from antarest.output.filestudy.file_output_utils import (
+from antarest.output.filestudy.logs import find_simulation_log
+from antarest.output.filestudy.metadata import (
     extract_output_details,
 )
-from antarest.output.filestudy.logs import find_simulation_log
-from antarest.output.filestudy.utils import QueryFileType
+from antarest.output.filestudy.model import QueryFileType
 from antarest.output.filestudy.variables import extract_variables_list
 from antarest.output.model import OutputVariablesList
 from antarest.output.model.download import MatrixIndex
